+++ flowauth/backend/flowauth/models.py
@@ -522,6 +522,7 @@ def make_demodata():  # pragma: no cover
         "geography",
         "unique_subscriber_counts",
         "location_introversion",
+        "total_network_objects",
     ):
         c = Capability(name=c)
         db.session.add(c)