+++ flowclient/flowclient/__init__.py
@@ -30,6 +30,7 @@ from .client import (
     get_available_dates,
     unique_subscriber_counts,
     location_introversion,
+    total_network_objects,
 )
 
 __all__ = [
@@ -52,4 +53,5 @@ __all__ = [
     "get_available_dates",
     "unique_subscriber_counts",
     "location_introversion",
+    "total_network_objects",
 ]