+++ flowclient/flowclient/client.py
@@ -981,7 +981,7 @@ def location_introversion(
         Unit of aggregation, e.g. "admin3"
     direction : {"in", "out", "both"}, default "both"
         Optionally, include only ingoing or outbound calls/texts
-
+>
     Returns
     -------
     dict
@@ -993,4 +993,32 @@ def location_introversion(
         "end_date": end_date,
         "aggregation_unit": aggregation_unit,
         "direction": direction,
+    }
+
+
+def total_network_objects(
+    start_date: str, end_date: str, aggregation_unit: str
+) -> dict:
+    """
+    Return query spec for total network objects
+
+    Parameters
+    ----------
+    start_date : str
+        ISO format date of the first day of the count, e.g. "2016-01-01"
+    end_date : str
+        ISO format date of the day _after_ the final date of the count, e.g. "2016-01-08"
+    aggregation_unit : str
+        Unit of aggregation, e.g. "admin3"
+
+    Returns
+    -------
+    dict
+        Dict which functions as the query specification
+    """
+    return {
+        "query_kind": "total_network_objects",
+        "start_date": start_date,
+        "end_date": end_date,
+        "aggregation_unit": aggregation_unit,
     }