+++ flowmachine/flowmachine/core/server/query_schemas/flowmachine_query.py
@@ -26,6 +26,7 @@ from .location_introversion import (
     LocationIntroversionSchema,
     LocationIntroversionExposed,
 )
+from .total_network_objects import TotalNetworkObjectsSchema, TotalNetworkObjectsExposed
 from .dfs_metric_total_amount import (
     DFSTotalMetricAmountSchema,
     DFSTotalMetricAmountExposed,
@@ -46,6 +47,7 @@ class FlowmachineQuerySchema(OneOfSchema):
         "location_event_counts": LocationEventCountsSchema,
         "unique_subscriber_counts": UniqueSubscriberCountsSchema,
         "location_introversion": LocationIntroversionSchema,
+        "total_network_objects": TotalNetworkObjectsSchema,
         "dfs_metric_total_amount": DFSTotalMetricAmountSchema,
     }
 
@@ -72,6 +74,8 @@ class FlowmachineQuerySchema(OneOfSchema):
             return "unique_subscriber_counts"
         elif isinstance(obj, LocationIntroversionExposed):
             return "location_introversion"
+        elif isinstance(obj, TotalNetworkObjectsExposed):
+            return "total_network_objects"
         elif isinstance(obj, DFSTotalMetricAmountExposed):
             return "dfs_metric_total_amount"
         else: