+++ integration_tests/tests/flowmachine_server_tests/test_server.py
@@ -54,6 +54,7 @@ def test_get_available_queries(send_zmq_message_and_receive_reply):
                 "location_event_counts",
                 "unique_subscriber_counts",
                 "location_introversion",
+                "total_network_objects",
                 "dfs_metric_total_amount",
             ]
         },