+++ integration_tests/tests/test_queries.py
@@ -35,6 +35,14 @@ from .utils import permissions_types, aggregation_types
             },
         ),
         (
+            "total_network_objects",
+            {
+                "start_date": "2016-01-01",
+                "end_date": "2016-01-02",
+                "aggregation_unit": "admin3",
+            },
+        ),
+        (
             "location_introversion",
             {
                 "start_date": "2016-01-01",